requirements.txt
python-gitlab
PyQt5